NP-1858 Move save function to actions file
